// src/app/features/issue/providers/kiami/kiami-api/kiami-api-responses.d.ts
export type KiamiOriginalIssueState = 'open' | 'closed' | 'all';
export type KiamiOriginalMilestoneState = 'active' | 'closed';
export type KiamiOriginalUserState = 'active' | 'blocked';

export type KiamiOriginalMilestone = Readonly<{
  id: number;
  iid: number;
  project_id: number;
  title: string;
  description: string;
  start_date: string;
  due_date: string;
  state: KiamiOriginalMilestoneState;
  created_at: string;
  updated_at: string;
}>;

export type KiamiOriginalUser = Readonly<{
  id: number;
  username: string;
  name: string;
  state: KiamiOriginalUserState;
  avatar_url: string;
  web_url: string;
}>;

export type KiamiOriginalIssue = Readonly<{
  id: number;
  iid: number;
  project_id: number;
  title: string;
  description: string;
  state: KiamiOriginalIssueState;
  weight: number;
  created_at: string;
  updated_at: string;
  closed_at: string;
  closed_by: string;
  labels: string[];
  milestone: unknown;
  assignees: KiamiOriginalUser[];
  author: KiamiOriginalUser;
  assignee: KiamiOriginalUser;
  user_notes_count: number;
  merge_requests_count: number;
  upvotes: number;
  downvotes: number;
  due_date: string;
  confidential: boolean;
  discussion_locked: boolean;
  web_url: string;
  time_stats: {
    time_estimate: number;
    total_time_spent: number;
    human_time_estimate: string;
    human_total_time_spent: string;
  };
  task_completion_status: {
    count: number;
    completed_count: number;
  };
  has_tasks: boolean;
  task_status: string;
  _links: {
    self: string;
    notes: string;
    award_emoji: string;
    project: string;
  };
  references: {
    short: string;
    relative: string;
    full: string;
  };
  moved_to_id: number;
}>;

export type KiamiOriginalComment = Readonly<{
  id: number;
  body: string;
  attachment: string;
  author: KiamiOriginalUser;
  created_at: string;
  updated_at: string;
  system: boolean;
  noteable_id: number;
  noteable_type: string;
  noteable_iid: number;
  resolvable: boolean;
}>;